Crate opentok

Modules

audio_device

connection

log

publisher

session

stream

subscriber

video_capturer

video_frame

Enums

OtcError

OpenTok error status codes.

Traits

IntoResult

Functions

deinit

Destroys the library engine. You should call this function when you are done executing code that uses the library.

init

Initializes the library. You must call this function before the execution of any other code using the library.

Type Definitions

OtcResult