Module opentok::audio_device

Structs

AudioDevice

AudioDeviceCallbacks

AudioDeviceCallbacksBuilder

AudioDeviceSettings

Settings for a AudioDevice.

AudioSample

High level storage for audio samples. The data size should correspond with the sampling_rate and number_of_channels (size = (sampling_rate / 1000) * number_of_channels).

AudioSampleData

Raw data holder for audio samples.

Enums

AudioDeviceError

Functions

initialize

Initialize the global AudioDevice. This is called immediately after the library initialization. The AudioDevice has no associated callback by default. The application call set_render_callbacks() and/or set_capture_callbacks() to have the AudioDevice configured.

set_capture_callbacks

Associate capture callbacks to the AudioDevice singleton. This should be called if the application wants to use a custom audio capturer.

set_render_callbacks

Associate render callbacks to the AudioDevice singleton. This should be called if the application wants to use a custom audio renderer.